aass_agents/requirements.txt
# Core ADK framework
google-adk>=1.27.4

# Search & HTTP
duckduckgo-search>=8.1.1
httpx>=0.28.1

# Data validation
pydantic>=2.12.5

# Environment
python-dotenv>=1.2.2

# Marketing tools
pytrends>=4.9.2
feedparser>=6.0.12

# API server
fastapi>=0.135.2
uvicorn>=0.42.0

# Browser automation
playwright>=1.44.0
playwright-stealth>=1.0.6
capsolver>=1.0.0

# YAML config (hook engine, tool registry)
pyyaml>=6.0.2

# Testing
pytest>=9.0.2
pytest-asyncio>=1.3.0
